+++ HISTORY.rst
@@ -158,6 +158,7 @@
 - The HOME_VIEW setting is not a Django setting but a setting
   from the common app. The HOME_VIEW has been moved to the
   COMMON namespace and renamed to COMMON_HOME_VIEW.
+- New link added to display the events of the current user.
 
 3.1.9 (2018-11-01)
 ==================